feat(web): add the Farm page as a living ASCII tulip field
